Get filepath from command line and offer to save results

+++ src/run_model.py
@@ -1,12 +1,14 @@
 import cPickle
+import sys
+import os.path
 
 import numpy as np
 from wordcloud import WordCloud
 
 import dataload
-from build_model import MODEL_FILEPATH
 
 PREDICTIONS_FILEPATH = "data/predictions.csv"
+WORDCLOUD_FILEPATH = "data/topwords.png"
 
 def get_top_words(model, n=50):
     """Return a list of the top n most important words found by the model"""
@@ -29,15 +31,35 @@ def get_top_words_cloud(featurenames, coefficients, num_words=100):
     wc.to_file('topwords.png')
 
 if __name__ == "__main__":
-    with open(MODEL_FILEPATH) as model_file:
+    if len(sys.argv) < 2:
+        sys.exit("USAGE: %s filepath-to-pickled-model" % sys.argv[0])
+
+    filepath = sys.argv[1]
+    if not os.path.isfile(filepath):
+        sys.exit("Error: File '%s' not found" % filepath)
+
+    with open(filepath) as model_file:
         model = cPickle.load(model_file)
-    df = dataload.get_aws_df()
+
+    df = dataload.get_labeled_df()
     df = df[df['in_genDR'] == 0]
-    
     corpus = df["GeneRIF text"]
 
     predictions = model.predict(corpus)
-
     df['dr_relevant'] = predictions
 
-    df.to_csv(PREDICTIONS_FILEPATH)
+    print "Top Words:"
+    top_words = get_top_words(model, n=100)
+    print ", ".join(top_words) 
+    
+    save_wordcloud = raw_input("Save wordcloud? (y/N)").lower()
+    if save_wordcloud == 'y':
+        coefs = model.named_steps['multinomialnb'].coef_[0]
+        featurenames = model.named_steps['tfidfvectorizer'].get_feature_names()
+        get_top_words_cloud(featurenames, coefs)
+        print "Saved as {0}".format(WORDCLOUD_FILEPATH)
+    save_csv = raw_input("Save csv? (y/N)").lower()
+    if save_csv == 'y':
+        df.to_csv(PREDICTIONS_FILEPATH)
+        print "Saved as {0}".format(PREDICTIONS_FILEPATH)
+
